separate cloud functions file
